refactor(recipe-view): Extract keywords to component.
